requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
alembic==1.16.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorlog==6.9.0
comm==0.2.3
contourpy==1.3.3
cryptography==44.0.3
cycler==0.12.1
datasets==3.1.0
debugpy==1.8.15
decorator==5.2.1
dill==0.3.8
executing==2.2.0
filelock==3.18.0
flwr==1.19.0
flwr-datasets==0.5.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.9.0
greenlet==3.2.3
grpcio==1.74.0
hf-xet==1.1.5
huggingface-hub==0.34.1
idna==3.10
imageio==2.37.0
ipykernel==6.30.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
iterators==0.0.2
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
optuna==4.4.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==4.25.8
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pycocotools==2.0.10
pycparser==2.22
pycryptodome==3.23.0
pydicom==3.0.1
Pygments==2.19.2
pylidc==0.2.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.31.0
referencing==0.36.2
regex==2025.7.31
requests==2.32.4
rich==13.9.4
rpds-py==0.26.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
seaborn==0.13.2
setuptools==78.1.1
shellingham==1.5.4
simpleitk==2.5.2
six==1.17.0
SQLAlchemy==2.0.41
stack-data==0.6.3
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.4
tomli==2.2.1
tomli_w==1.2.0
torch==2.5.1
torchvision==0.20.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.1
triton==3.1.0
typer==0.12.5
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.171
ultralytics-thop==2.0.14
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
